checkboxes_v0/monetization_prop/protyping_api_web_call/twilio_app/requirements.txt
pyjwt
flask==2.3.3
twilio==9.3.0
python-dotenv==1.0.1
flask-cors==4.0.1